fix: Combine slowmode and typing indicators into one ui
