+++ packages/client/components/app/interface/channels/text/Messages.tsx
@@ -66,11 +66,6 @@ interface Props {
   pendingMessages?: (props: { tail: boolean; ids: string[] }) => JSX.Element;
 
   /**
-   * Display typing indicator instead of padding
-   */
-  typingIndicator?: JSX.Element;
-
-  /**
    * Highlighted message id
    */
   highlightedMessageId: Accessor<string | undefined>;
@@ -959,7 +954,6 @@ export function Messages(props: Props) {
                   tail: pendingMessageIsTrailing(),
                   ids: sentMessageIdempotency(),
                 })}
-                {props.typingIndicator ?? <Padding />}
               </Show>
             </div>
           </div>
@@ -987,15 +981,6 @@ const AnchorToEnd = styled("div", {
       position: "absolute",
       bottom: "var(--gap-md)",
     },
-  },
-});
-
-/**
- * Container padding
- */
-const Padding = styled("div", {
-  base: {
-    height: "24px",
   },
 });
