+++ packages/client/components/ui/components/features/messaging/composition/index.ts
@@ -2,4 +2,3 @@ export { FileCarousel } from "./FileCarousel";
 export { MessageBox } from "./MessageBox";
 export { MessageReplyPreview } from "./MessageReplyPreview";
 export * from "./picker";
-export { TypingIndicator } from "./TypingIndicator";
