+++ packages/client/src/interface/channels/text/Composition.tsx
@@ -1,9 +1,6 @@
-import { createCountdownFromNow } from "@solid-primitives/date";
 import {
   For,
-  Match,
   Show,
-  Switch,
   createEffect,
   createMemo,
   createSignal,
@@ -14,11 +11,8 @@ import {
 import { useLingui } from "@lingui/solid/macro";
 import { Channel } from "stoat.js";
 
-import { styled } from "styled-system/jsx";
-
 import { useClient } from "@revolt/client";
 import { debounce } from "@revolt/common";
-import { useDurationFormat } from "@revolt/i18n/durations";
 import { useInstance } from "@revolt/instance";
 import { Keybind, KeybindAction, createKeybind } from "@revolt/keybinds";
 import { useModals } from "@revolt/modal";
@@ -31,7 +25,6 @@ import {
   IconButton,
   MessageBox,
   MessageReplyPreview,
-  Tooltip,
   humanFileSize,
 } from "@revolt/ui";
 import { Symbol } from "@revolt/ui/components/utils/Symbol";
@@ -58,61 +51,6 @@ export function MessageComposition(props: Props) {
   const client = useClient();
   const { limits } = useInstance();
   const { openModal } = useModals();
-  const durationFormat = useDurationFormat();
-
-  const isSlowmodeExempt = (): boolean => {
-    return props.channel.havePermission("BypassSlowmode");
-  };
-
-  const slowmodeCountdown = createMemo(() => {
-    if (!props.channel.slowmode || isSlowmodeExempt()) return 0;
-
-    const entry = props.channel.userSlowmode();
-    if (!entry) return;
-
-    const receivedAt = entry.receivedAt ?? Date.now();
-    // Add 100 ms here so the countdown has a bit to render
-    const targetTs = receivedAt + 100 + entry.retry_after * 1000;
-    return createCountdownFromNow(targetTs);
-  });
-
-  const cooldownRemaining = createMemo(() => {
-    const cd = slowmodeCountdown();
-
-    if (!cd) return 0;
-
-    const [store] = cd;
-
-    const h = store.hours ?? 0;
-    const m = store.minutes ?? 0;
-    const s = store.seconds ?? 0;
-
-    const totalSeconds = h * 3600 + m * 60 + s;
-    return totalSeconds > 0 ? totalSeconds : 0;
-  });
-
-  const slowmodeText = createMemo(() => {
-    const cd = slowmodeCountdown();
-
-    if (!cd) return "";
-
-    const [store] = cd;
-
-    return durationFormat(
-      { seconds: store.seconds, minutes: store.minutes, hours: store.hours },
-      { style: "digital" },
-    );
-  });
-
-  const slowmodeWaitTime = createMemo(() => {
-    const s = props.channel.slowmode;
-    if (!s) return "";
-    const h = Math.floor(s / 3600);
-    const m = Math.floor((s % 3600) / 60);
-    const sec = s % 60;
-
-    return durationFormat({ seconds: sec, minutes: m, hours: h });
-  });
 
   createKeybind(KeybindAction.CHAT_JUMP_END, () =>
     setNodeReplacement(["_focus"]),
@@ -371,24 +309,6 @@ export function MessageComposition(props: Props) {
 
   return (
     <>
-      <Show when={props.channel.slowmode}>
-        <SlowmodeContainer>
-          <Tooltip
-            content={t`Members can send one message every ${slowmodeWaitTime()}.`}
-            placement="top"
-          >
-            <SlowmodeRow>
-              <Symbol style={{ "font-size": "1rem" }}>schedule</Symbol>
-              <SlowmodeText>
-                <Switch fallback={t`Slowmode is enabled.`}>
-                  <Match when={isSlowmodeExempt()}>{t`Slowmode Immune`}</Match>
-                  <Match when={cooldownRemaining() > 0}>{slowmodeText()}</Match>
-                </Switch>
-              </SlowmodeText>
-            </SlowmodeRow>
-          </Tooltip>
-        </SlowmodeContainer>
-      </Show>
       <Show when={state.draft.hasAdditionalElements(props.channel.id)}>
         <Keybind
           keybind={KeybindAction.CHAT_REMOVE_COMPOSITION_ELEMENT}
@@ -525,26 +445,3 @@ export function MessageComposition(props: Props) {
     </>
   );
 }
-
-const SlowmodeContainer = styled("div", {
-  base: {
-    display: "flex",
-    justifyContent: "flex-end",
-    padding: "0 12px 6px 0",
-  },
-});
-
-const SlowmodeRow = styled("div", {
-  base: {
-    display: "flex",
-    alignItems: "center",
-    gap: "var(--gap-sm)",
-  },
-});
-
-const SlowmodeText = styled("span", {
-  base: {
-    fontSize: "0.75rem",
-    fontWeight: "600",
-  },
-});
